fixing compile errors in MessageBroker
moving MessageBroker to its own file
