+++ Assets/Mirage/Runtime/MessageBroker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 58b5dd2e149643745b982057ab9ecd9c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
